requirements.txt
aiogram==2.25.1
aiohttp==3.8.1
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
frozendict==2.3.4
frozenlist==1.3.3
html5lib==1.1
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.2
pandas==1.5.3
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
webencodings==0.5.1
yarl==1.8.2
